screen_scanner/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
pyautogui>=0.9.52
keyboard>=0.13.5